don't set `chat_parallel(max_active)` and `rpm` (closes #66)
...and allow for setting by the user.

I do feel that it's pretty unpleasant to hit the rate limit by default rather than seeing responses more slowly-but-surely trickling in, but this seems to be mostly a Claude-specific effect and many users aren't likely to encounter this. If they do, the fix probably lives most cozily in ellmer.